Batch shell (#46)
* feat: Add accounts to the address book from existing seeds (#88)

The `new_account` command generates a random secret seed for the new alias. This
patch adds a `-s` command line switch that lets a user specify a secret seed for
the account.

* feat: Add load_batch command (#89)

This adds a command to the interactive shell that runs the commands in the
specified file as if they were typed into the repl. The '#' character may be
used create a comment that extends to the end of the line.

* fixup! feat: Add load_batch command (#89)